backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
PyMuPDF>=1.23.0
beautifulsoup4>=4.12.0
python-docx>=1.1.0
anthropic>=0.20.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
